+++ docs/assets/navigation.js
@@ -1,0 +1,1 @@
+window.navigationData = "eJyLrlYqSa0oUbJScjIKcM7JTM0rUdJRKkgsyVCyUkrOSSwuTi3Wh0vpZZTk5ijpKGVn5qUoWRkaWdTGAgC6DBUE"