+++ src/types/card-types.ts
@@ -1,0 +1,7 @@
+export interface TokenizeCardParams {
+    cardHolderName: string;
+    cardNumber: string;
+    cardExpirationDate: string;
+    cardCvv: string;
+    cardBrand: string;
+}